Added depth buffer, borders, respawn, timer (lessons 40-43)

+++ scr/Renderer/SpriteAnimator.cpp
@@ -9,7 +9,7 @@ namespace RenderEngine {
         , m_currentFrameDuration(m_pSprite->getFrameDuration(0))
         , m_currentAnimationTime(0) {}
 
-    void SpriteAnimator::update(const uint64_t delta) {
+    void SpriteAnimator::update(const double delta) {
         m_currentAnimationTime += delta;
         while (m_currentAnimationTime >= m_currentFrameDuration) {
             m_currentAnimationTime -= m_currentFrameDuration;
